refactor: rename managers
