+++ src/main/java/app/SessionManager.java
@@ -1,4 +1,4 @@
-package session;
+package app;
 
 import java.util.HashMap;
 import java.util.Map;
